IntegralRepresentationMixin: new interface for objects
This Mixin class is for functions or other objects that can be written
in terms of integrals.  For example if they have definitions in terms
of integrals.  Classes need only implement an `_as_integral` to have
working `.rewrite(Integral)` and `.as_integral(...)`, with default
variable of integration and appropriate documentation.
